API-12087 remove common objects from agents
