+++ src/agent/rtm.ts
@@ -1,31 +1,35 @@
 import type {
-  ListChatsResponse,
-  ListChatParameters,
+  AgentForTransfer,
+  ChangePushNotificationsRequest,
+  CreateCustomerResponse,
+  CustomerParameters,
   EmptyResponse,
-  ListThreadsParameters,
-  ListThreadsResponse,
+  Event,
   GetChatResponse,
+  GetCustomerResponse,
   ListArchivesParameters,
   ListArchivesResponse,
-  StartChatParameters,
-  StartChatResponse,
+  ListChatParameters,
+  ListChatsResponse,
+  ListThreadsParameters,
+  ListThreadsResponse,
+  LoginRequest,
+  LoginResponse,
+  MulticastRecipients,
+  Properties,
+  Push,
+  Pushes,
   ResumeChatParameters,
   ResumeChatResponse,
-  TransferChatParameters,
+  RoutingStatus,
+  RTMAPIOptions,
   SendEventResponse,
   SendRichMessagePostbackParameters,
-  GetCustomerResponse,
-  CustomerParameters,
-  CreateCustomerResponse,
-  MulticastRecipients,
-  AgentForTransfer,
-  LoginResponse,
-  ChangePushNotificationsRequest,
-  LoginRequest,
-  Pushes,
   SetRoutingStatusResponse,
+  StartChatParameters,
+  StartChatResponse,
+  TransferChatParameters,
 } from "./structures";
-import { Event, Properties, Push, RoutingStatus, RTMAPIOptions } from "../objects";
 import { RTMAPI } from "../internal";
 
 export default class RTM extends RTMAPI {
